refactor: create background.js
